[#2801] Remove V1 server map builder
