+++ web/src/main/java/com/navercorp/pinpoint/web/applicationmap/ApplicationMapBuilderFactory.java
@@ -19,10 +19,7 @@ package com.navercorp.pinpoint.web.applicationmap;
 import com.navercorp.pinpoint.web.applicationmap.appender.histogram.NodeHistogramAppenderFactory;
 import com.navercorp.pinpoint.web.applicationmap.appender.server.ServerInfoAppenderFactory;
 import com.navercorp.pinpoint.web.vo.Range;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 /**
@@ -31,29 +28,19 @@ import org.springframework.stereotype.Component;
 @Component
 public class ApplicationMapBuilderFactory {
 
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    private final String mode;
-
     private final NodeHistogramAppenderFactory nodeHistogramAppenderFactory;
 
     private final ServerInfoAppenderFactory serverInfoAppenderFactory;
 
     @Autowired
     public ApplicationMapBuilderFactory(
-            @Value("#{pinpointWebProps['web.servermap.builder.mode'] ?: 'v1'}") String mode,
             NodeHistogramAppenderFactory nodeHistogramAppenderFactory,
             ServerInfoAppenderFactory serverInfoAppenderFactory) {
-        this.mode = mode;
         this.nodeHistogramAppenderFactory = nodeHistogramAppenderFactory;
         this.serverInfoAppenderFactory = serverInfoAppenderFactory;
     }
 
     public ApplicationMapBuilder createApplicationMapBuilder(Range range) {
-        logger.info("ApplicationMapBuilder mode : {}", mode);
-        if (mode.equals("v2")) {
-            return new ApplicationMapBuilderV2(range, nodeHistogramAppenderFactory, serverInfoAppenderFactory);
-        }
-        return new ApplicationMapBuilderV1(range);
+        return new ApplicationMapBuilder(range, nodeHistogramAppenderFactory, serverInfoAppenderFactory);
     }
 }
