+++ web/src/main/java/com/navercorp/pinpoint/web/service/map/LinkSelectorFactory.java
@@ -16,13 +16,9 @@
 
 package com.navercorp.pinpoint.web.service.map;
 
-import com.navercorp.pinpoint.web.dao.HostApplicationMapDao;
 import com.navercorp.pinpoint.web.security.ServerMapDataFilter;
 import com.navercorp.pinpoint.web.service.LinkDataMapService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 /**
@@ -31,13 +27,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class LinkSelectorFactory {
 
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    private final String mode;
-
     private final LinkDataMapService linkDataMapService;
 
-    private final HostApplicationMapDao hostApplicationMapDao;
 
     private final ApplicationsMapCreatorFactory applicationsMapCreatorFactory;
 
@@ -45,35 +36,25 @@ public class LinkSelectorFactory {
 
     @Autowired(required = false)
     public LinkSelectorFactory(
-            @Value("#{pinpointWebProps['web.servermap.selector.mode'] ?: 'v1'}") String mode,
             LinkDataMapService linkDataMapService,
-            HostApplicationMapDao hostApplicationMapDao,
             ApplicationsMapCreatorFactory appliationsMapCreatorFactory) {
-        this(mode, linkDataMapService, hostApplicationMapDao, appliationsMapCreatorFactory, null);
+        this(linkDataMapService, appliationsMapCreatorFactory, null);
     }
 
     @Autowired(required = false)
     public LinkSelectorFactory(
-            @Value("#{pinpointWebProps['web.servermap.selector.mode'] ?: 'v1'}") String mode,
             LinkDataMapService linkDataMapService,
-            HostApplicationMapDao hostApplicationMapDao,
             ApplicationsMapCreatorFactory appliationsMapCreatorFactory,
             ServerMapDataFilter serverMapDataFilter) {
-        logger.info("LinkSelector mode : {}", mode);
-        this.mode = mode;
         this.linkDataMapService = linkDataMapService;
-        this.hostApplicationMapDao = hostApplicationMapDao;
         this.applicationsMapCreatorFactory = appliationsMapCreatorFactory;
         this.serverMapDataFilter = serverMapDataFilter;
     }
 
     public LinkSelector create() {
-        if (mode.equalsIgnoreCase("v2")) {
-            VirtualLinkMarker virtualLinkMarker = new VirtualLinkMarker();
-            VirtualLinkProcessor virtualLinkProcessor = new VirtualLinkProcessor(linkDataMapService, virtualLinkMarker);
-            ApplicationsMapCreator applicationsMapCreator = applicationsMapCreatorFactory.create(virtualLinkMarker);
-            return new BFSLinkSelectorV2(applicationsMapCreator, virtualLinkProcessor, serverMapDataFilter);
-        }
-        return new BFSLinkSelector(linkDataMapService, hostApplicationMapDao, serverMapDataFilter);
+        VirtualLinkMarker virtualLinkMarker = new VirtualLinkMarker();
+        VirtualLinkProcessor virtualLinkProcessor = new VirtualLinkProcessor(linkDataMapService, virtualLinkMarker);
+        ApplicationsMapCreator applicationsMapCreator = applicationsMapCreatorFactory.create(virtualLinkMarker);
+        return new BFSLinkSelector(applicationsMapCreator, virtualLinkProcessor, serverMapDataFilter);
     }
 }
