+++ web/src/main/resources/pinpoint-web.properties
@@ -45,18 +45,10 @@ web.hbase.selectAllSpans.limit=500
 
 web.activethread.activeAgent.duration.days=7
 
-# server map link selector mode = v1 or v2 (default = v1)
-web.servermap.selector.mode=v2
-
 # server map link creator mode = serial or parallel (default = serial)
-# only applicable when web.servermap.selector.mode=v2
 web.servermap.creator.mode=parallel
 web.servermap.creator.parallel.maxthreads=16
 
-# server map builder mode = v1 or v2 (default = v1)
-web.servermap.builder.mode=v2
-
 # server map appender mode = serial or parallel (default = serial)
-# only applicable when web.servermap.builder.mode=v2
 web.servermap.appender.mode=parallel
 web.servermap.appender.parallel.maxthreads=16
