--- web/src/test/java/com/navercorp/pinpoint/web/service/map/BFSLinkSelectorTest.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017 NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.navercorp.pinpoint.web.service.map;
-
-import com.navercorp.pinpoint.web.dao.HostApplicationMapDao;
-import com.navercorp.pinpoint.web.service.LinkDataMapService;
-
-/**
- * @author emeroad
- */
-public class BFSLinkSelectorTest extends LinkSelectorTestBase {
-
-    @Override
-    protected LinkSelectorFactory createLinkSelectorFactory(LinkDataMapService linkDataMapService, HostApplicationMapDao hostApplicationMapDao) {
-        return new LinkSelectorFactory("v1", linkDataMapService, hostApplicationMapDao, null);
-    }
-}
-
