+++ web/src/test/java/com/navercorp/pinpoint/web/service/map/BFSLinkSelector_parallel_Test.java
@@ -22,11 +22,10 @@ import com.navercorp.pinpoint.web.service.LinkDataMapService;
 /**
  * @author HyunGil Jeong
  */
-public class BFSLinkSelectorV2_parallel_Test extends LinkSelectorTestBase {
+public class BFSLinkSelector_parallel_Test extends LinkSelectorTestBase {
 
     @Override
-    protected LinkSelectorFactory createLinkSelectorFactory(LinkDataMapService linkDataMapService, HostApplicationMapDao hostApplicationMapDao) {
-        ApplicationsMapCreatorFactory applicationsMapCreatorFactory = new ApplicationsMapCreatorFactory("parallel", 16, hostApplicationMapDao, linkDataMapService);
-        return new LinkSelectorFactory("v2", linkDataMapService, hostApplicationMapDao, applicationsMapCreatorFactory);
+    protected ApplicationsMapCreatorFactory createApplicationsMapCreatorFactory(LinkDataMapService linkDataMapService, HostApplicationMapDao hostApplicationMapDao) {
+        return new ApplicationsMapCreatorFactory("parallel", 16, hostApplicationMapDao, linkDataMapService);
     }
 }
